requirements.txt
attrs==21.4.0
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
Flask-JWT==0.3.2
Flask-SQLAlchemy==2.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
SQLAlchemy==1.4.36
tomli==2.0.1
Werkzeug==2.1.2
